+++ src/node/rpc/endpoint.h
@@ -4,6 +4,7 @@
 
 #include "ds/json.h"
 #include "http/authentication/authentication_types.h"
+#include "kv/serialise_entry_blit.h"
 #include "node/service_map.h"
 
 #include <string>
@@ -52,6 +53,37 @@ namespace ccf
 MSGPACK_ADD_ENUM(ccf::endpoints::ForwardingRequired);
 MSGPACK_ADD_ENUM(ccf::endpoints::ExecuteOutsideConsensus);
 MSGPACK_ADD_ENUM(ccf::endpoints::Mode);
+
+namespace kv::serialisers
+{
+  template <>
+  struct BlitSerialiser<ccf::endpoints::EndpointKey>
+  {
+    static SerialisedEntry to_serialised(
+      const ccf::endpoints::EndpointKey& endpoint_key)
+    {
+      size_t size_ = sizeof(size_t) + endpoint_key.uri_path.size() +
+        sizeof(endpoint_key.verb);
+      SerialisedEntry data(size_);
+      auto data_ = data.data();
+
+      serialized::write(data_, size_, endpoint_key.uri_path);
+      serialized::write(data_, size_, endpoint_key.verb);
+      return data;
+    }
+
+    static ccf::endpoints::EndpointKey from_serialised(
+      const SerialisedEntry& data)
+    {
+      auto data_ = data.data();
+      auto size_ = data.size();
+
+      auto uri_path = serialized::read<ccf::endpoints::URI>(data_, size_);
+      auto verb = serialized::read<ccf::RESTVerb>(data_, size_);
+      return {uri_path, verb};
+    }
+  };
+}
 
 namespace ccf
 {
